refactor(roadflare-rider): extract RideTab state-binding components
Complete the deferred portion of #67 held back in PR #68.

The six stage-specific composables (IdleContent, RequestingContent,
ChoosingDriverContent, MatchedContent, InRideContent, CompletedContent)
were left in RideTab.kt because they bound to types (RideSession,
DriverInfo, ChatMessage, RouteResult, RiderViewModel) whose packages
were being stabilised in PR #65. Those packages are now stable on main,
so these composables move to components/RideTabComponents.kt alongside
CancelledContent and formatFareAmount.

RideTab.kt trims from 766 → 133 lines: just the tab scaffold and the
when-routing that dispatches to each state component. RideTabComponents.kt
grows from 54 → 678 lines holding all eight composables plus the fare
formatter. Each component takes narrow props plus modifier = Modifier;
IdleContent keeps a RiderViewModel parameter because it wires ~20
state flows and actions — a narrow-props rewrite is a separate concern.

Drops ChoosingDriverContent's unused rideReferenceFareUsd parameter
and the matching caller arg. Removes two stale imports from RideTab.kt
(AdminConfig, Location). No behavior changes.
